prima: Adapt WLAN_MSG_WLAN_TP_IND netlink message.
In old prima drivers, the WLAN_MSG_WLAN_TP_IND hint
is sent as an simple integer value, which defines the throughput level.
In newer prima drivers ( qcacld-3.0 ) this is extended by an flag,
where the enable_tcp_delack and enable_tcp_adv_win_scale parameters
can be added.
This is hint is processed by the cnss-daemon which adjusts then
tcp kernel values according these hints.

Newer cnss-daemons expect this structure and not a simple integer.
This patch uses the structures from qcacld-3.0 and applies them for the
prima driver which should be used in conjunction with newer cnss-daemons.